feat: Restructure and update Policy Abstractions
This commit refactors the policy abstractions to improve logical organization, enhance consistency, and prepare for future enhancements.

***

### Key Changes:

-   **Namespace Restructuring**:
    -   The `IPolicyContext` interface was moved from `src/Policies/IPolicyContext.cs` to a new `Contexts` subdirectory, resulting in the new path `src/Policies/Contexts/IPolicyContext.cs`. Its namespace was also updated to `Zentient.Abstractions.Policies.Contexts` to reflect this change.
    -   The `IPolicy<out TPolicyDefinition>` interface was also moved out of the `Definitions` namespace and into the base `Policies` namespace to better reflect its role as the executable policy abstraction, not just a definition. The updated file is `src/Policies/IPolicy{out TPolicyDefinition}.cs` and the namespace is now `Zentient.Abstractions.Policies`.

-   **Dependency Updates**:
    -   `IPolicyContext` now has a `using Zentient.Abstractions.Policies.Definitions;` directive to reference the policy definition type.
    -   `IPolicy<out TPolicyDefinition>` now has `using Zentient.Abstractions.Policies.Contexts;` and `using Zentient.Abstractions.Policies.Definitions;` to correctly reference its dependencies.

-   **API Enhancements**:
    -   The `IPolicyDescriptor` interface has been updated to include a `using Zentient.Abstractions.Metadata.Readers;` directive. This change prepares for a future where policy descriptors use the more specific, read-only `IMetadataReader` interface, which provides a safer and more consistent way to access metadata. This is a non-functional change at the interface level, but it signals the intent to use the new metadata reading abstraction.

These changes are aimed at creating a more coherent and maintainable policy system within the framework. The new namespace structure separates the core policy logic from its definitions and contexts, making the purpose of each component clearer. The dependency and API updates align the policy abstractions with recent framework-wide changes to metadata handling.
