+++ src/Policies/IPolicyDescriptor.cs
@@ -4,6 +4,7 @@
 
 using Zentient.Abstractions.Common;
 using Zentient.Abstractions.Metadata;
+using Zentient.Abstractions.Metadata.Readers;
 
 namespace Zentient.Abstractions.Policies
 {
