+++ src/Policies/IPolicy{out TPolicyDefinition}.cs
@@ -2,7 +2,10 @@
 // Copyright © 2025 Zentient Framework Team. All rights reserved.
 // </copyright>
 
-namespace Zentient.Abstractions.Policies.Definitions
+using Zentient.Abstractions.Policies.Contexts;
+using Zentient.Abstractions.Policies.Definitions;
+
+namespace Zentient.Abstractions.Policies
 {
     /// <summary>
     /// Represents a policy that can be applied to an operation producing a result of type <typeparamref name="TPolicyDefinition"/>.
